pyproject.toml
[project]
name = "monitoramento-impressoras-agente"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel R",email = "gabrielrodriguesalves1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "pysnmp-lextudio (==6.1.2)",
    "aioping (>=0.4.0,<0.5.0)",
    "pyasn1 (==0.5.1)",
    "monitoramento-impressoras-core @ git+ssh://git@github.com/gabriel-r-alves/monitoramento_impressoras_core.git",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.typos.files]
extend-exclude = ["*.md"]

[tool.taskipy.tasks]
pre_lint = 'typos'
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev monitoramento_impressoras_agente/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=monitoramento_impressoras_agente -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.10,<0.16.0)",
    "typos (>=1.45.0,<2.0.0)"
]
